+++ sdk/build-scripts/package.json
@@ -9,8 +9,7 @@
 		"node": ">=16"
 	},
 	"bin": {
-		"build-package": "./src/build-package.ts",
-		"build-dapp-kit": "./src/build-dapp-kit.ts"
+		"build-package": "./src/build-package.ts"
 	},
 	"scripts": {
 		"prettier:check": "prettier -c --ignore-unknown .",
@@ -22,10 +21,7 @@
 		"build": "tsc --build"
 	},
 	"devDependencies": {
-		"@vanilla-extract/esbuild-plugin": "^2.3.0",
-		"autoprefixer": "^10.4.14",
-		"esbuild": "^0.17.6",
-		"postcss": "^8.4.26",
+		"esbuild": "^0.18.14",
 		"typescript": "^5.1.6"
 	},
 	"dependencies": {
