--- sdk/dapp-kit/src/components/ConnectButton.css.ts
@@ -1,8 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-import { style } from '@vanilla-extract/css';
-
-export const testStyle = style({
-	padding: 10,
-});
