--- sdk/dapp-kit/src/components/ConnectButton.tsx
@@ -1,8 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-import * as styles from './ConnectButton.css.js';
-
-export function ConnectButton() {
-	return <div className={styles.testStyle}>test</div>;
-}
