+++ sdk/dapp-kit/src/index.ts
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export * from './hooks/useSuiClient.js';
-export * from './components/ConnectButton.js';
 export * from './components/SuiClientProvider.js';
 export * from './components/WalletProvider.js';
 export * from './hooks/useRpcApiVersion.js';
