+++ sdk/dapp-kit/vitest.config.ts
@@ -2,11 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import tsconfigPaths from 'vite-tsconfig-paths';
-import { vanillaExtractPlugin } from '@vanilla-extract/vite-plugin';
 import { defineConfig, configDefaults } from 'vitest/config';
 
 export default defineConfig({
-	plugins: [tsconfigPaths(), vanillaExtractPlugin()],
+	plugins: [tsconfigPaths()],
 	test: {
 		exclude: [...configDefaults.exclude, 'tests/**'],
 		environment: 'happy-dom',
